Id Fase a Long

+++ src/main/java/com/pucp/aevent/dao/IFaseDao.java
@@ -12,7 +12,7 @@ public interface IFaseDao  extends JpaRepository <Fase, Long>{
 	
 	public List<Fase> findByIdEvento(Evento evento);
 	
-	public Fase findByIdFase(Integer idFase);
+	public Fase findByIdFase(Long idFase);
 	
-	public Fase deleteByIdFase(Integer idFase);
+	public Fase deleteByIdFase(Long idFase);
 }
